Debug Button, and Health System
In this small little commit I added the Health System and the Debug
Button. The debug button functionality will change but the key for it is
numpad / if you wanna give it a go. Health System needs some tweaking
but for now when you have a certain amount of food and thirst it will
add health. Might change or remove in the future still don't know. Next
commit I'm going to clean up some code and add more comments.